+++ src/AuditaX.EntityFramework/Customizers/AuditaXModelCustomizer.cs
@@ -61,12 +61,12 @@ internal sealed class AuditaXModelCustomizer : ModelCustomizer
 
             entity.Property(e => e.SourceName)
                 .HasColumnName(sourceNameColumn)
-                .HasMaxLength(50)
+                .HasMaxLength(64)
                 .IsRequired();
 
             entity.Property(e => e.SourceKey)
                 .HasColumnName(sourceKeyColumn)
-                .HasMaxLength(900)
+                .HasMaxLength(64)
                 .IsRequired();
 
             entity.Property(e => e.AuditLogXml)